requirements.txt
pandas
numpy
gspread
requests
dash
google-auth
gunicorn